refactor(chat): extract chat list primitives
