+++ src/components/chat-list-mobile-rail-port.test.ts
@@ -3,6 +3,7 @@ import assert from "node:assert/strict";
 import { readFileSync } from "node:fs";
 
 const source = readFileSync(new URL("./chat-list.tsx", import.meta.url), "utf8");
+const primitives = readFileSync(new URL("./chat-list-primitives.tsx", import.meta.url), "utf8");
 const styles = readFileSync(new URL("../app/globals.css", import.meta.url), "utf8");
 const topBar = readFileSync(new URL("./top-bar.tsx", import.meta.url), "utf8");
 const workspace = readFileSync(new URL("./workspace.tsx", import.meta.url), "utf8");
@@ -16,9 +17,9 @@ assert.match(topBar, /labeled=\{familiarSwitcherLabeled\}/, "TopBar can render t
 assert.match(workspace, /familiarSwitcherLabeled=\{mode === "chat"\}/, "Workspace labels the top-bar selector on the Familiars page");
 
 // ── Counted PINNED / SESSIONS section headers ────────────────────────────────
-assert.match(source, /function ChatListSection/, "ChatList gains a counted section-header primitive");
+assert.match(primitives, /function ChatListSection/, "ChatList gains a counted section-header primitive");
 assert.match(
-  source,
+  primitives,
   /uppercase tracking-\[0\.12em\]/,
   "Section headers are uppercase + letter-spaced like the desktop rail",
 );
